fix circular dep in eth package
