+++ packages/eth/src/services/Config.ts
@@ -6,22 +6,9 @@
  * as some factory function to generate configs for various providers.
  */
 
-import { chains } from "./Wallet.js"
-import { createConfig, fallback, http } from "@wagmi/core"
+import { chains, getTransportWithFallback } from "./Wallet.js"
+import { createConfig } from "@wagmi/core"
 import { transports } from "./_unstable.js"
-import { config } from "@fxhash/config"
-import { Chain } from "viem"
-
-export const getTransportWithFallback = (chain: Chain) => {
-  return fallback([
-    http(),
-    http(
-      chain.id === chains.ETHEREUM.id
-        ? config.eth.apis.alchemy.rpc
-        : config.base.apis.alchemy.rpc
-    ),
-  ])
-}
 
 export const supportedEvmChains = [chains.ETHEREUM, chains.BASE] as const
 export const viemTransports = [transports.ETHEREUM, transports.BASE] as const
